Rename coingecko wrapper file
