+++ freqtrade/util/coin_gecko.py
